feat: 게시물 개인 작성자 삭제 가능

+++ backend/src/main/java/com/example/backend/controller/board/BoardController.java
@@ -4,6 +4,7 @@ import com.example.backend.dto.board.Board;
 import com.example.backend.service.board.BoardService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
@@ -35,15 +36,24 @@ public class BoardController {
     }
 
     @DeleteMapping("boardDelete/{boardId}")
-    public ResponseEntity<Map<String, Object>> delete(@PathVariable int boardId) {
-        if (service.remove(boardId)) {
-            return ResponseEntity.ok()
-                    .body(Map.of("message", Map.of("type", "success"
-                            , "text", STR."\{boardId}번 게시글이 삭제되었습니다.")));
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity<Map<String, Object>> delete(
+            @PathVariable int boardId,
+            Authentication authentication) {
+        if (service.hasAccess(boardId, authentication)) {
+            if (service.remove(boardId)) {
+                return ResponseEntity.ok()
+                        .body(Map.of("message", Map.of("type", "success"
+                                , "text", STR."\{boardId}번 게시글이 삭제되었습니다.")));
+            } else {
+                return ResponseEntity.internalServerError()
+                        .body(Map.of("message", Map.of("type", "error"
+                                , "text", "게시글 삭제 중 문제가 발생하였습니다.")));
+            }
         } else {
-            return ResponseEntity.internalServerError()
+            return ResponseEntity.status(403)
                     .body(Map.of("message", Map.of("type", "error"
-                            , "text", "게시글 삭제 중 문제가 발생하였습니다.")));
+                            , "text", "삭제 권한이 없습니다.")));
         }
     }
